pyproject.toml
[project]
name = "university-pythonweb-hw-12"
version = "0.1.0"
description = ""
authors = [
    {name = "KashnykovAnton",email = "kashnykov.anton@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "sqlalchemy (==2.0.31)",
    "alembic (==1.13.2)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "fastapi (>=0.115.11,<0.116.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "greenlet (>=3.1.1,<4.0.0)",
    "redis (>=5.2.1,<6.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "apscheduler (>=3.11.0,<4.0.0)",
    "slowapi (>=0.1.9,<0.2.0)",
    "fastapi-mail (>=1.4.2,<2.0.0)",
    "libgravatar (>=1.0.4,<2.0.0)",
    "cloudinary (>=1.43.0,<2.0.0)",
    "uvicorn (>=0.29.0,<0.30.0)",
    "python-multipart>=0.0.6,<1.0.0",
    "psycopg2-binary>=2.9.9",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
pytest = "^8.3.5"
pytest-asyncio = "^0.26.0"
aiosqlite = "^0.21.0"
pytest-cov = "^6.0.0"
sphinx = "^8.2.3"

[tool.pytest.ini_options]
addopts = "--doctest-modules -W ignore::DeprecationWarning -W ignore::UserWarning"
testpaths = ["tests"]
pythonpath = "."
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning"
]
asyncio_default_fixture_loop_scope = "function"
